backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai==1.58.1
qdrant-client==1.12.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
langchain==0.3.14
langchain-openai==0.2.14
langchain-qdrant==0.1.4
httpx==0.28.1
